chore: Optimize code and add utils in the Ang Component
