+++ components/Ang/index.tsx
@@ -1,8 +1,8 @@
 import React, {useEffect, useState} from 'react';
-import {Text, View} from 'react-native';
+import {View} from 'react-native';
 import {AngProps, AngData} from './interfaces';
-import {textStyles} from '../../styles/text';
 import {layoutStyles} from '../../styles/layout';
+import {bake} from './utils/bakePankti';
 
 const Ang = (props: AngProps): JSX.Element => {
   const [currentAngData, setCurrentAngData] = useState({} as AngData);
@@ -13,39 +13,12 @@ const Ang = (props: AngProps): JSX.Element => {
       .then(data => setCurrentAngData(data));
   }, [props.page]);
 
-  const getStyle = (index: number) => {
-    if (props.larivaarAssist) {
-      if (index % 2 === 0) {
-        return {
-          color: 'black',
-          ...textStyles.wordStyle,
-        };
-      } else {
-        return {
-          color: 'red',
-          ...textStyles.wordStyle,
-        };
-      }
-    } else {
-      return textStyles.wordStyle;
-    }
-  };
-
-  const bake = (verse: string) => {
-    const words = verse.split(' ');
-    return words.map((word, index) => {
-      return (
-        <View key={index}>
-          <Text style={getStyle(index)}>{word}</Text>
-        </View>
-      );
-    });
-  };
-
   return (
     <View style={layoutStyles.wordContainer}>
       {currentAngData.page &&
-        currentAngData.page.map(page => bake(page.verse.unicode))}
+        currentAngData.page.map(page =>
+          bake(page.verse.unicode, props.larivaarAssist),
+        )}
     </View>
   );
 };
