+++ styles/elements.tsx
@@ -1,0 +1,13 @@
+import {StyleSheet} from 'react-native';
+
+export const elementStyles = StyleSheet.create({
+  gurbani: {
+    fontSize: 20,
+  },
+  input: {
+    borderStyle: 'solid',
+    borderWidth: 1,
+    borderColor: 'black',
+    padding: 10,
+  },
+});
