--- styles/text.tsx
@@ -1,7 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export const textStyles = StyleSheet.create({
-  wordStyle: {
-    fontSize: 20,
-  },
-});
